requirements.txt
websockets>=12.0
RPi.GPIO>=0.7.1
